refactor(tui): extract composer history search
Move the Ctrl-R composer search session, footer rendering, and match highlight helpers into a child module so chat_composer.rs owns less feature-specific state.

Keep traversal and dedupe in chat_composer_history.rs while moving the focused search tests next to the extracted implementation.
